pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "odmcp"
dynamic = ["version"]
description = "Python MCP servers exposing Open Data to compatible LLM clients."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.0.1",
]

[tool.hatch.version]
path = "src/odmcp/__init__.py"

[project.scripts]
odmcp = "odmcp.cli:main"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.pyright]
include = ["src/mcp", "tests"]
venvPath = "."
venv = ".venv"

[dependency-groups]
dev = [
    "pytest-asyncio>=0.24.0",
    "pytest>=8.3.3",
    "ruff>=0.8.1",
    "pre-commit>=4.0.1",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-ra -q"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
# Configure log format
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"


[tool.hatch.build.targets.wheel]
packages = ["src/odmcp"]
